`iface.ValidatorDB`: Implement `SlashableProposalCheck`.
