+++ validator/client/BUILD.bazel
@@ -61,8 +61,6 @@ go_library(
         "//validator/client/validator-client-factory:go_default_library",
         "//validator/db:go_default_library",
         "//validator/db/common:go_default_library",
-        "//validator/db/filesystem:go_default_library",
-        "//validator/db/kv:go_default_library",
         "//validator/graffiti:go_default_library",
         "//validator/helpers:go_default_library",
         "//validator/keymanager:go_default_library",
@@ -104,7 +102,6 @@ go_test(
         "attest_test.go",
         "key_reload_test.go",
         "metrics_test.go",
-        "propose_protect_test.go",
         "propose_test.go",
         "registration_test.go",
         "runner_test.go",
