+++ validator/db/iface/BUILD.bazel
@@ -12,6 +12,7 @@ go_library(
     deps = [
         "//config/fieldparams:go_default_library",
         "//config/proposer:go_default_library",
+        "//consensus-types/interfaces:go_default_library",
         "//consensus-types/primitives:go_default_library",
         "//monitoring/backup:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
