+++ validator/db/kv/proposer_protection.go
@@ -5,8 +5,10 @@ import (
 	"fmt"
 
 	"github.com/pkg/errors"
+	"github.com/prometheus/client_golang/prometheus"
 	fieldparams "github.com/prysmaticlabs/prysm/v5/config/fieldparams"
 	"github.com/prysmaticlabs/prysm/v5/config/params"
+	"github.com/prysmaticlabs/prysm/v5/consensus-types/interfaces"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/primitives"
 	"github.com/prysmaticlabs/prysm/v5/encoding/bytesutil"
 	"github.com/prysmaticlabs/prysm/v5/time/slots"
@@ -191,6 +193,86 @@ func (s *Store) HighestSignedProposal(ctx context.Context, publicKey [fieldparam
 		return nil
 	})
 	return highestSignedProposalSlot, exists, err
+}
+
+// SlashableProposalCheck checks if a block proposal is slashable by comparing it with the
+// block proposals history for the given public key in our complete slashing protection database defined by EIP-3076.
+// If it is not, we then update the history.
+func (s *Store) SlashableProposalCheck(
+	ctx context.Context,
+	pubKey [fieldparams.BLSPubkeyLength]byte,
+	signedBlock interfaces.ReadOnlySignedBeaconBlock,
+	signingRoot [32]byte,
+	emitAccountMetrics bool,
+	validatorProposeFailVec *prometheus.CounterVec,
+) error {
+	fmtKey := fmt.Sprintf("%#x", pubKey[:])
+
+	blk := signedBlock.Block()
+	prevSigningRoot, proposalAtSlotExists, prevSigningRootExists, err := s.ProposalHistoryForSlot(ctx, pubKey, blk.Slot())
+	if err != nil {
+		if emitAccountMetrics {
+			validatorProposeFailVec.WithLabelValues(fmtKey).Inc()
+		}
+		return errors.Wrap(err, "failed to get proposal history")
+	}
+
+	lowestSignedProposalSlot, lowestProposalExists, err := s.LowestSignedProposal(ctx, pubKey)
+	if err != nil {
+		return err
+	}
+
+	// Based on EIP-3076 - Condition 2
+	// -------------------------------
+	if lowestProposalExists {
+		// If the block slot is (strictly) less than the lowest signed proposal slot in the DB, we consider it slashable.
+		if blk.Slot() < lowestSignedProposalSlot {
+			return fmt.Errorf(
+				"could not sign block with slot < lowest signed slot in db, block slot: %d < lowest signed slot: %d",
+				blk.Slot(),
+				lowestSignedProposalSlot,
+			)
+		}
+
+		// If the block slot is equal to the lowest signed proposal slot and
+		// - condition1: there is no signed proposal in the DB for this slot, or
+		// - condition2: there is  a signed proposal in the DB for this slot, but with no associated signing root, or
+		// - condition3: there is  a signed proposal in the DB for this slot, but the signing root differs,
+		// ==> we consider it slashable.
+		condition1 := !proposalAtSlotExists
+		condition2 := proposalAtSlotExists && !prevSigningRootExists
+		condition3 := proposalAtSlotExists && prevSigningRootExists && prevSigningRoot != signingRoot
+		if blk.Slot() == lowestSignedProposalSlot && (condition1 || condition2 || condition3) {
+			return fmt.Errorf(
+				"could not sign block with slot == lowest signed slot in db if it is not a repeat signing, block slot: %d == slowest signed slot: %d",
+				blk.Slot(),
+				lowestSignedProposalSlot,
+			)
+		}
+	}
+
+	// Based on EIP-3076 - Condition 1
+	// -------------------------------
+	// If there is a signed proposal in the DB for this slot and
+	// - there is no associated signing root, or
+	// - the signing root differs,
+	// ==> we consider it slashable.
+	if proposalAtSlotExists && (!prevSigningRootExists || prevSigningRoot != signingRoot) {
+		if emitAccountMetrics {
+			validatorProposeFailVec.WithLabelValues(fmtKey).Inc()
+		}
+		return errors.New(common.FailedBlockSignLocalErr)
+	}
+
+	// Save the proposal for this slot.
+	if err := s.SaveProposalHistoryForSlot(ctx, pubKey, blk.Slot(), signingRoot[:]); err != nil {
+		if emitAccountMetrics {
+			validatorProposeFailVec.WithLabelValues(fmtKey).Inc()
+		}
+		return errors.Wrap(err, "failed to save updated proposal history")
+	}
+
+	return nil
 }
 
 func pruneProposalHistoryBySlot(valBucket *bolt.Bucket, newestSlot primitives.Slot) error {
