+++ validator/db/kv/proposer_protection_test.go
@@ -4,12 +4,16 @@ import (
 	"context"
 	"testing"
 
+	"github.com/ethereum/go-ethereum/common/hexutil"
 	fieldparams "github.com/prysmaticlabs/prysm/v5/config/fieldparams"
 	"github.com/prysmaticlabs/prysm/v5/config/params"
+	"github.com/prysmaticlabs/prysm/v5/consensus-types/blocks"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/primitives"
 	"github.com/prysmaticlabs/prysm/v5/encoding/bytesutil"
+	ethpb "github.com/prysmaticlabs/prysm/v5/proto/prysm/v1alpha1"
 	"github.com/prysmaticlabs/prysm/v5/testing/assert"
 	"github.com/prysmaticlabs/prysm/v5/testing/require"
+	"github.com/prysmaticlabs/prysm/v5/testing/util"
 	"github.com/prysmaticlabs/prysm/v5/validator/db/common"
 )
 
@@ -300,4 +304,152 @@ func TestStore_HighestSignedProposal(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, true, exists)
 	assert.Equal(t, primitives.Slot(3), slot)
+}
+
+func Test_slashableProposalCheck_PreventsLowerThanMinProposal(t *testing.T) {
+	ctx := context.Background()
+	lowestSignedSlot := primitives.Slot(10)
+
+	var pubkey [fieldparams.BLSPubkeyLength]byte
+	pubkeyBytes, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
+	require.NoError(t, err, "Failed to decode pubkey")
+	copy(pubkey[:], pubkeyBytes)
+
+	db := setupDB(t, [][fieldparams.BLSPubkeyLength]byte{pubkey})
+	require.NoError(t, err)
+
+	// We save a proposal at the lowest signed slot in the DB.
+	err = db.SaveProposalHistoryForSlot(ctx, pubkey, lowestSignedSlot, []byte{1})
+	require.NoError(t, err)
+
+	// We expect the same block with a slot lower than the lowest
+	// signed slot to fail validation.
+	blk := &ethpb.SignedBeaconBlock{
+		Block: &ethpb.BeaconBlock{
+			Slot:          lowestSignedSlot - 1,
+			ProposerIndex: 0,
+			Body:          &ethpb.BeaconBlockBody{},
+		},
+		Signature: params.BeaconConfig().EmptySignature[:],
+	}
+	wsb, err := blocks.NewSignedBeaconBlock(blk)
+	require.NoError(t, err)
+	err = db.SlashableProposalCheck(context.Background(), pubkey, wsb, [32]byte{4}, false, nil)
+	require.ErrorContains(t, "could not sign block with slot < lowest signed", err)
+
+	// We expect the same block with a slot equal to the lowest
+	// signed slot to pass validation if signing roots are equal.
+	blk = &ethpb.SignedBeaconBlock{
+		Block: &ethpb.BeaconBlock{
+			Slot:          lowestSignedSlot,
+			ProposerIndex: 0,
+			Body:          &ethpb.BeaconBlockBody{},
+		},
+		Signature: params.BeaconConfig().EmptySignature[:],
+	}
+	wsb, err = blocks.NewSignedBeaconBlock(blk)
+	require.NoError(t, err)
+	err = db.SlashableProposalCheck(ctx, pubkey, wsb, [32]byte{1}, false, nil)
+	require.NoError(t, err)
+
+	// We expect the same block with a slot equal to the lowest
+	// signed slot to fail validation if signing roots are different.
+	wsb, err = blocks.NewSignedBeaconBlock(blk)
+	require.NoError(t, err)
+	err = db.SlashableProposalCheck(ctx, pubkey, wsb, [32]byte{4}, false, nil)
+	require.ErrorContains(t, "could not sign block with slot == lowest signed", err)
+
+	// We expect the same block with a slot > than the lowest
+	// signed slot to pass validation.
+	blk = &ethpb.SignedBeaconBlock{
+		Block: &ethpb.BeaconBlock{
+			Slot:          lowestSignedSlot + 1,
+			ProposerIndex: 0,
+			Body:          &ethpb.BeaconBlockBody{},
+		},
+		Signature: params.BeaconConfig().EmptySignature[:],
+	}
+
+	wsb, err = blocks.NewSignedBeaconBlock(blk)
+	require.NoError(t, err)
+	err = db.SlashableProposalCheck(ctx, pubkey, wsb, [32]byte{3}, false, nil)
+	require.NoError(t, err)
+}
+
+func Test_slashableProposalCheck(t *testing.T) {
+	ctx := context.Background()
+
+	var pubkey [fieldparams.BLSPubkeyLength]byte
+	pubkeyBytes, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
+	require.NoError(t, err, "Failed to decode pubkey")
+	copy(pubkey[:], pubkeyBytes)
+
+	db := setupDB(t, [][fieldparams.BLSPubkeyLength]byte{pubkey})
+	require.NoError(t, err)
+
+	blk := util.HydrateSignedBeaconBlock(&ethpb.SignedBeaconBlock{
+		Block: &ethpb.BeaconBlock{
+			Slot:          10,
+			ProposerIndex: 0,
+			Body:          &ethpb.BeaconBlockBody{},
+		},
+		Signature: params.BeaconConfig().EmptySignature[:],
+	})
+
+	// We save a proposal at slot 1 as our lowest proposal.
+	err = db.SaveProposalHistoryForSlot(ctx, pubkey, 1, []byte{1})
+	require.NoError(t, err)
+
+	// We save a proposal at slot 10 with a dummy signing root.
+	dummySigningRoot := [32]byte{1}
+	err = db.SaveProposalHistoryForSlot(ctx, pubkey, 10, dummySigningRoot[:])
+	require.NoError(t, err)
+	sBlock, err := blocks.NewSignedBeaconBlock(blk)
+	require.NoError(t, err)
+
+	err = db.SlashableProposalCheck(ctx, pubkey, sBlock, dummySigningRoot, false, nil)
+	// We expect the same block sent out with the same root should not be slasahble.
+	require.NoError(t, err)
+
+	// We expect the same block sent out with a different signing root should be slashable.
+	err = db.SlashableProposalCheck(ctx, pubkey, sBlock, [32]byte{2}, false, nil)
+	require.ErrorContains(t, common.FailedBlockSignLocalErr, err)
+
+	// We save a proposal at slot 11 with a nil signing root.
+	blk.Block.Slot = 11
+	sBlock, err = blocks.NewSignedBeaconBlock(blk)
+	require.NoError(t, err)
+	err = db.SaveProposalHistoryForSlot(ctx, pubkey, blk.Block.Slot, nil)
+	require.NoError(t, err)
+
+	// We expect the same block sent out should return slashable error even
+	// if we had a nil signing root stored in the database.
+	err = db.SlashableProposalCheck(ctx, pubkey, sBlock, [32]byte{2}, false, nil)
+	require.ErrorContains(t, common.FailedBlockSignLocalErr, err)
+
+	// A block with a different slot for which we do not have a proposing history
+	// should not be failing validation.
+	blk.Block.Slot = 9
+	sBlock, err = blocks.NewSignedBeaconBlock(blk)
+	require.NoError(t, err)
+	err = db.SlashableProposalCheck(ctx, pubkey, sBlock, [32]byte{3}, false, nil)
+	require.NoError(t, err, "Expected allowed block not to throw error")
+}
+
+func Test_slashableProposalCheck_RemoteProtection(t *testing.T) {
+	var pubkey [fieldparams.BLSPubkeyLength]byte
+	pubkeyBytes, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
+	require.NoError(t, err, "Failed to decode pubkey")
+	copy(pubkey[:], pubkeyBytes)
+
+	db := setupDB(t, [][fieldparams.BLSPubkeyLength]byte{pubkey})
+	require.NoError(t, err)
+
+	blk := util.NewBeaconBlock()
+	blk.Block.Slot = 10
+	sBlock, err := blocks.NewSignedBeaconBlock(blk)
+	require.NoError(t, err)
+
+	err = db.SlashableProposalCheck(context.Background(), pubkey, sBlock, [32]byte{2}, false, nil)
+	require.NoError(t, err, "Expected allowed block not to throw error")
 }
