+++ validator/helpers/BUILD.bazel
@@ -31,6 +31,7 @@ go_test(
     deps = [
         "//config/fieldparams:go_default_library",
         "//config/proposer:go_default_library",
+        "//consensus-types/interfaces:go_default_library",
         "//consensus-types/primitives:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
         "//testing/require:go_default_library",
